added validation
