+++ backend/requirements.txt
@@ -28,6 +28,7 @@ pyflakes==2.2.0
 PyJWT==1.7.1
 pyparsing==2.4.7
 pytest==6.1.1
+pytest-asyncio==0.14.0
 pytest-cov==2.10.1
 python-dateutil==2.8.1
 python-dotenv==0.15.0
@@ -40,4 +41,5 @@ toml==0.10.1
 typing-extensions==3.7.4.3
 urllib3==1.26.4
 uvicorn==0.12.2
+validate-email==1.3
 zipp==3.3.2
